workers/file-saver/requirements.txt
pydantic>=2.0.0,<3.0.0
python-magic>=0.4.27,<0.5.0
loguru>=0.7.0,<0.8.0